requirements.txt
scikit-learn
skforecast
pandas
numpy
supabase
python-dotenv
pyyaml
matplotlib
seaborn
xgboost
lightgbm
